Expand Developing with Webpack
Now this starts to feel like a whole chapter. I discussed loaders in
more detail here though that needs more info on later chapters. Also
discussed briefly about how to manage configuration as it begins to
grow. I decided to put the React chapter after this one because it feels
like it could build on top of it naturally.