Push binds to constructors where it makes sense
Now it's more consistent and we get rid of that `() =>` weirdness. It's
preferable to bind at constructor due to performance. Also `render`
methods look simpler this way. Later on this can be refactored to even
simpler form if `::` makes it to the standard or a stable way to
autobind appears.

Closes #106.