building kanban - Document a better hashing approach
The problem with the default one (require CSS chunk) is that it
invalidates JS/CSS if either belonging to the same chunk changes. It's
more sensible to define a separate entry point for CSS in this case.

Perhaps this would be a good default for the book. You see `require`
style a lot, but this style has its merits. Only gotcha is that Webpack
generates an extra JS file for the entry point.